+++ src/sensor.cpp
@@ -1514,7 +1514,7 @@ namespace librealsense
             auto&& composite = dynamic_cast<composite_frame*>(f.frame);
             if (composite)
             {
-                for (int i = 0; i < composite->get_embedded_frames_count(); i++)
+                for (size_t i = 0; i < composite->get_embedded_frames_count(); i++)
                 {
                     processed_frames.push_back(composite->get_frame(i));
                 }
